Update MessageBuilder Tests
